refactor: avoid prisma in client components
