--- src/actions.ts
@@ -1,9 +1,0 @@
-'use server';
-
-import { signOut } from '@/auth';
-// import { redirect } from 'next/navigation';
-
-export async function signOutAction() {
-  return await signOut({ redirectTo: '/' });
-}
-
