+++ src/app/candidate/browse/page.tsx
@@ -1,9 +1,6 @@
 import DashboardClient from "../../../components/DashboardClient";
-import {
-  getFilterOptions,
-  FilterConfig,
-  ActiveFilters,
-} from "../../../app/api/filterOptions";
+import { getFilterOptions } from "../../../app/api/filterOptions";
+import type { FilterConfig, ActiveFilters } from "@/lib/filters";
 import { listUsers } from "../../../app/api/users/list";
 import { Role } from "@prisma/client";
